// libs/client/files/feature/files-table/src/lib/one-drive-picker/picker-sdk/types.ts
export interface IAuthenticateCommand {
  command: 'authenticate';
  resource: string;
  type: 'SharePoint' | 'Graph';
  claimsChallenge?: {
    claims: string;
    token?: string;
    error: any; // IError;
  };
  expiration?: {
    error: any; // IError;
  };
}